requirements.txt
discord.py>=2.3.0
psutil>=5.9.0